refactor(core): move the index/, sparse_index/ and quantization/ inline test modules to sibling files (#1918)
Second of four lots for velesdb-core. Twelve files, and the lot where the
audit's awkward shapes are concentrated: only three are straight moves.

- wal_framing.rs holds a pub(crate) mod io_counters, not a private mod
  tests. The extraction tool deliberately refuses this shape rather than
  emit a wrong declaration, so it was moved by hand: the sibling keeps the
  body, and the declaration keeps both pub(crate) and the twenty-six line
  doc comment explaining why the counters are thread-local. Two other test
  files reach these counters through crate::index::wal_framing::io_counters,
  so losing the visibility would have broken them.
- trigram/gpu.rs carries two blocks under different gates: mod tests under
  #[cfg(test)] and mod gpu_tests under #[cfg(all(test, feature = "gpu"))].
  They land in gpu_tests.rs and gpu_feature_tests.rs, each keeping its own
  gate and module name. Verified with cargo check --features gpu, since the
  default build does not compile the second block.
- locking.rs keeps its #[cfg(all(test, debug_assertions))] gate verbatim.
- Six files name their module something other than tests (safety_tests,
  probe_tests, load_bound_tests, offset_bound_tests, codec_helpers_tests,
  storage_mode_parsing_tests); the names are kept and the siblings named
  after them. search.rs and persistence.rs also already had a <stem>_tests.rs
  sibling, which the module name sidesteps naturally.

The sorted test-name list is byte-identical before and after, re-checked
after rustfmt: 5357 tests, 5341 passing and 16 pre-existing ignored.
Baseline down to 69 files.
