+++ scripts/inline-tests-baseline.txt
@@ -42,28 +42,16 @@ crates/velesdb-core/src/gpu/gpu_csr.rs	141
 crates/velesdb-core/src/gpu/gpu_traversal.rs	229
 crates/velesdb-core/src/gpu/pq_gpu.rs	119
 crates/velesdb-core/src/half_precision.rs	73
-crates/velesdb-core/src/index/hnsw/index/mod.rs	18
-crates/velesdb-core/src/index/hnsw/native/graph/locking.rs	44
-crates/velesdb-core/src/index/hnsw/native/graph/reorder.rs	100
-crates/velesdb-core/src/index/hnsw/native/graph/search.rs	95
-crates/velesdb-core/src/index/hnsw/native/graph_io.rs	46
-crates/velesdb-core/src/index/sparse/persistence.rs	26
-crates/velesdb-core/src/index/trigram/gpu.rs	114
-crates/velesdb-core/src/index/wal_framing.rs	91
 crates/velesdb-core/src/metrics/guardrails.rs	100
 crates/velesdb-core/src/metrics/latency.rs	38
 crates/velesdb-core/src/metrics/operational.rs	97
 crates/velesdb-core/src/metrics/query.rs	108
 crates/velesdb-core/src/metrics/retrieval.rs	76
-crates/velesdb-core/src/quantization/codec_helpers.rs	62
-crates/velesdb-core/src/quantization/mod.rs	101
 crates/velesdb-core/src/scored_result.rs	57
 crates/velesdb-core/src/simd_dispatch.rs	276
 crates/velesdb-core/src/simd_native/adc.rs	118
 crates/velesdb-core/src/simd_neon.rs	77
 crates/velesdb-core/src/simd_neon_prefetch.rs	53
-crates/velesdb-core/src/sparse_index/search/mod.rs	243
-crates/velesdb-core/src/sparse_index/types.rs	110
 crates/velesdb-core/src/storage/async_ops.rs	35
 crates/velesdb-core/src/storage/atomic_write.rs	57
 crates/velesdb-core/src/test_fixtures.rs	51
